Fix rabbitmqbox build issues
- Rename running/restart_box to avoid conflict with gw/shared.h
- Update deprecated amqp.h includes to rabbitmq-c/*
- Fix octstr_hex_to_binary usage (in-place conversion)
- Fix cfg_get_integer type (int -> long for port/heartbeat/prefetch)
- Add .gitignore for build artifacts

+++ addons/rabbitmqbox/gw/rmq_connection.h
@@ -9,8 +9,8 @@
 #define RMQ_CONNECTION_H
 
 #include "gwlib/gwlib.h"
-#include <amqp.h>
-#include <amqp_tcp_socket.h>
+#include <rabbitmq-c/amqp.h>
+#include <rabbitmq-c/tcp_socket.h>
 
 /* RabbitMQ connection state */
 typedef struct {
@@ -20,12 +20,12 @@ typedef struct {
 
     /* Connection parameters */
     Octstr *host;
-    int port;
+    long port;
     Octstr *vhost;
     Octstr *username;
     Octstr *password;
-    int heartbeat;
-    int prefetch;
+    long heartbeat;
+    long prefetch;
 
     /* TLS/SSL */
     int use_ssl;
